legacy_archive/backup_before_ipados/requirements.txt
PySide6
qt-material
img2pdf
Pillow
PySide2
